backend/requirements.txt
# Core RAG dependencies
pymupdf>=1.23.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
numpy>=1.24.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0